Simplify loading for statistics and maps

+++ WarriorsSnuggery/Statistics.cs
@@ -1,6 +1,6 @@
 using System.Collections.Generic;
 using System.IO;
-using System.Linq;
+using WarriorsSnuggery.Loader;
 using WarriorsSnuggery.Maps;
 using WarriorsSnuggery.Objects.Parts;
 
@@ -9,21 +9,21 @@ namespace WarriorsSnuggery
 	public sealed class GameStatistics
 	{
 		// Paths
-		public string Name { get; private set; }
-		public string SaveName { get; private set; }
+		public string Name;
+		public string SaveName;
 		// Changing Values
 		public int Level;
 		public int Money;
 		public string Actor;
-		public float RelativeHP;
+		public float Health;
 		public int Mana;
 		public int Kills;
 		public int Deaths;
 		public int MaxMana;
 
-		public GameMode Mode;
-		public GameType Type;
-		public string MapType;
+		public GameMode CurrentMode;
+		public GameType CurrentType;
+		public string CurrentMapType;
 		public int Waves;
 		public List<bool[]> Shroud;
 
@@ -51,22 +51,22 @@ namespace WarriorsSnuggery
 			Level = save.Level;
 			Money = save.Money;
 			Actor = save.Actor;
-			RelativeHP = save.RelativeHP;
+			Health = save.Health;
 			Mana = save.Mana;
 			Kills = save.Kills;
 			Deaths = save.Deaths;
 
-			Mode = save.Mode;
-			Type = save.Type;
-			MapType = save.MapType;
+			CurrentMode = save.CurrentMode;
+			CurrentType = save.CurrentType;
+			CurrentMapType = save.CurrentMapType;
 			Waves = save.Waves;
 			Shroud = save.Shroud;
 
-			SpellCasters = save.SpellCasters;
+			SpellCasters = new Dictionary<int, (float, float)>(save.SpellCasters);
 
-			UnlockedSpells = save.UnlockedSpells.ToList();
-			UnlockedActors = save.UnlockedActors.ToList();
-			UnlockedTrophies = save.UnlockedTrophies.ToList();
+			UnlockedSpells = new List<string>(save.UnlockedSpells);
+			UnlockedActors = new List<string>(save.UnlockedActors);
+			UnlockedTrophies = new List<string>(save.UnlockedTrophies);
 
 			FinalLevel = save.FinalLevel;
 			Difficulty = save.Difficulty;
@@ -78,19 +78,96 @@ namespace WarriorsSnuggery
 			ScriptValues = save.ScriptValues;
 		}
 
-		GameStatistics() { }
+		public GameStatistics(string file)
+		{
+			SaveName = file;
+
+			var fields = PartLoader.GetFields(this, false);
+
+			foreach (var node in RuleReader.FindAndRead(FileExplorer.Saves, file, ".yaml"))
+			{
+				switch (node.Key)
+				{
+					case "Shroud":
+						Shroud = new List<bool[]>();
+
+						foreach (var node2 in node.Children)
+							Shroud.Add(node2.Convert<bool[]>());
+
+						break;
+					case "SpellCasters":
+
+						foreach (var node2 in node.Children)
+						{
+							var id = node2.Convert<int>();
+
+							var recharge = 0f;
+							var duration = 0f;
+							foreach (var node3 in node2.Children)
+							{
+								switch (node3.Key)
+								{
+									case "Recharge":
+										recharge = node3.Convert<float>();
+
+										break;
+									case "Remaining":
+										duration = node3.Convert<float>();
+
+										break;
+								}
+							}
+
+							SpellCasters.Add(id, (duration, recharge));
+						}
+
+						break;
+					case "UnlockedSpells":
+						foreach (var node2 in node.Children)
+							UnlockedSpells.Add(node2.Key);
+
+						break;
+					case "UnlockedActors":
+						foreach (var node2 in node.Children)
+							UnlockedActors.Add(node2.Key);
+
+						break;
+					case "UnlockedTrophies":
+						foreach (var node2 in node.Children)
+							UnlockedTrophies.Add(node2.Key);
+
+						break;
+					case "Script":
+						Script = node.Convert<string>();
+						ScriptValues = node.Children.ToArray();
+
+						break;
+					default:
+						PartLoader.SetValue(this, fields, node);
+
+						break;
+				}
+			}
+		}
+
+		public GameStatistics(int difficulty, bool hardcore, string name, int seed)
+		{
+			SetName(name);
+
+			Hardcore = hardcore;
+			Difficulty = difficulty;
+
+			Level = 1;
+			FinalLevel = (difficulty + 1) * 5;
+			Money = 100 - difficulty * 10;
+			MaxMana = GameSaveManager.DefaultStatistic.MaxMana;
+			Actor = GameSaveManager.DefaultStatistic.Actor;
+			Seed = seed;
+		}
 
 		public GameStatistics Copy()
 		{
 			return new GameStatistics(this);
-		}
-
-		public void Update(Game game)
-		{
-			if (game.World.PlayerSwitching)
-				RelativeHP = ((PlayerSwitchPart)game.World.LocalPlayer.Parts.Find(p => p is PlayerSwitchPart)).RelativeHP;
-			else
-				RelativeHP = game.World.LocalPlayer.Health == null ? 1 : game.World.LocalPlayer.Health.RelativeHP;
 		}
 
 		public bool ActorAvailable(PlayablePartInfo playable)
@@ -103,7 +180,7 @@ namespace WarriorsSnuggery
 			// Positive Points
 			var score = Level * 100 / FinalLevel;
 			score += Kills * 5;
-			score += Mana * 2;
+			score += Money * 2;
 			// Negative Points
 			score -= Deaths * 25;
 			return score;
@@ -113,9 +190,14 @@ namespace WarriorsSnuggery
 		{
 			var scriptState = world.Game.GetScriptState(out Script);
 
-			Mode = world.Game.Mode;
-			Type = world.Game.Type;
-			MapType = MapCreator.GetName(world.Map.Type, world.Game.Statistics);
+			if (world.PlayerSwitching)
+				Health = ((PlayerSwitchPart)world.LocalPlayer.Parts.Find(p => p is PlayerSwitchPart)).RelativeHP;
+			else
+				Health = world.LocalPlayer.Health == null ? 1 : world.LocalPlayer.Health.RelativeHP;
+
+			CurrentMode = world.Game.Mode;
+			CurrentType = world.Game.Type;
+			CurrentMapType = MapCreator.GetName(world.Map.Type, world.Game.Statistics);
 			Waves = world.Game.CurrentWave();
 
 			using (var writer = new StreamWriter(FileExplorer.Saves + SaveName + ".yaml", false))
@@ -129,9 +211,9 @@ namespace WarriorsSnuggery
 				writer.WriteLine("MaxMana=" + MaxMana);
 				writer.WriteLine("Kills=" + Kills);
 				writer.WriteLine("Deaths=" + Deaths);
-				writer.WriteLine("CurrentMode=" + Mode);
-				writer.WriteLine("CurrentType=" + Type);
-				writer.WriteLine("CurrentMapType=" + MapType);
+				writer.WriteLine("CurrentMode=" + CurrentMode);
+				writer.WriteLine("CurrentType=" + CurrentType);
+				writer.WriteLine("CurrentMapType=" + CurrentMapType);
 				if (Waves != 0)
 					writer.WriteLine("Waves=" + Waves);
 
@@ -156,13 +238,16 @@ namespace WarriorsSnuggery
 				writer.WriteLine("Seed=" + Seed);
 				writer.WriteLine("Mana=" + Mana);
 				writer.WriteLine("Actor=" + Actor);
-				writer.WriteLine("\tHealth=" + RelativeHP);
+				writer.WriteLine("\tHealth=" + Health);
+
 				writer.WriteLine("UnlockedSpells=");
 				foreach (var unlock in UnlockedSpells)
 					writer.WriteLine("\t" + unlock + "=");
+
 				writer.WriteLine("UnlockedActors=");
 				foreach (var unlock in UnlockedActors)
 					writer.WriteLine("\t" + unlock + "=");
+
 				writer.WriteLine("UnlockedTrophies=");
 				foreach (var unlock in UnlockedTrophies)
 					writer.WriteLine("\t" + unlock + "=");
@@ -184,7 +269,8 @@ namespace WarriorsSnuggery
 
 		public void Delete()
 		{
-			File.Delete(FileExplorer.Saves + SaveName + ".yaml");
+			if (File.Exists(FileExplorer.Saves + SaveName + ".yaml"))
+				File.Delete(FileExplorer.Saves + SaveName + ".yaml");
 
 			if (File.Exists(FileExplorer.Saves + SaveName + "_map.yaml"))
 				File.Delete(FileExplorer.Saves + SaveName + "_map.yaml");
@@ -198,153 +284,6 @@ namespace WarriorsSnuggery
 				name = name.Replace(c, '-');
 
 			SaveName = name;
-		}
-
-		public static GameStatistics CreateGameStatistic(int difficulty, bool hardcore, string name, int seed)
-		{
-			var statistic = new GameStatistics
-			{
-				Name = name,
-				Hardcore = hardcore,
-				Difficulty = difficulty
-			};
-			statistic.SetName(name);
-
-			statistic.Level = 1;
-			statistic.FinalLevel = (difficulty + 1) * 5;
-			statistic.Money = 100 - difficulty * 10;
-			statistic.MaxMana = GameSaveManager.DefaultStatistic.MaxMana;
-			statistic.Actor = GameSaveManager.DefaultStatistic.Actor;
-			statistic.Seed = seed;
-
-			return statistic;
-		}
-
-		public static GameStatistics LoadGameStatistic(string name)
-		{
-			var statistic = new GameStatistics
-			{
-				SaveName = name
-			};
-
-			foreach (var node in RuleReader.FindAndRead(FileExplorer.Saves, name, ".yaml"))
-			{
-				switch (node.Key)
-				{
-					case "Name":
-						statistic.Name = node.Convert<string>();
-						break;
-					case "Level":
-						statistic.Level = node.Convert<int>();
-						break;
-					case "Difficulty":
-						statistic.Difficulty = node.Convert<int>();
-						break;
-					case "Hardcore":
-						statistic.Hardcore = node.Convert<bool>();
-						break;
-					case "Money":
-						statistic.Money = node.Convert<int>();
-						break;
-					case "LevelAim":
-						statistic.FinalLevel = node.Convert<int>();
-						break;
-					case "MaxMana":
-						statistic.MaxMana = node.Convert<int>();
-						break;
-					case "Kills":
-						statistic.Kills = node.Convert<int>();
-						break;
-					case "Deaths":
-						statistic.Deaths = node.Convert<int>();
-						break;
-					case "CurrentMode":
-						statistic.Mode = node.Convert<GameMode>();
-						break;
-					case "CurrentType":
-						statistic.Type = node.Convert<GameType>();
-						break;
-					case "CurrentMapType":
-						statistic.MapType = node.Value;
-						break;
-					case "Waves":
-						statistic.Waves = node.Convert<int>();
-						break;
-					case "Shroud":
-						statistic.Shroud = new List<bool[]>();
-
-						foreach (var node2 in node.Children)
-							statistic.Shroud.Add(node2.Convert<bool[]>());
-
-						break;
-					case "Seed":
-						statistic.Seed = node.Convert<int>();
-						break;
-					case "Mana":
-						statistic.Mana = node.Convert<int>();
-						break;
-					case "Actor":
-						statistic.Actor = node.Convert<string>();
-
-						foreach (var node2 in node.Children)
-						{
-							if (node2.Key == "Health")
-								statistic.RelativeHP = node2.Convert<float>();
-							else
-								throw new YamlUnknownNodeException(node2.Key, name + ".yaml");
-						}
-						break;
-					case "SpellCasters":
-
-						foreach (var node2 in node.Children)
-						{
-							var id = node2.Convert<int>();
-
-							var recharge = 0f;
-							var duration = 0f;
-							foreach (var node3 in node2.Children)
-							{
-								switch(node3.Key)
-								{
-									case "Recharge":
-										recharge = node3.Convert<float>();
-
-										break;
-									case "Remaining":
-										duration = node3.Convert<float>();
-
-										break;
-								}
-							}
-
-							statistic.SpellCasters.Add(id, (duration, recharge));
-						}
-
-						break;
-					case "UnlockedSpells":
-						foreach (var node2 in node.Children)
-							statistic.UnlockedSpells.Add(node2.Key);
-
-						break;
-					case "UnlockedActors":
-						foreach (var node2 in node.Children)
-							statistic.UnlockedActors.Add(node2.Key);
-
-						break;
-					case "UnlockedTrophies":
-						foreach (var node2 in node.Children)
-							statistic.UnlockedTrophies.Add(node2.Key);
-
-						break;
-					case "Script":
-						statistic.Script = node.Value;
-						statistic.ScriptValues = node.Children.ToArray();
-
-						break;
-				}
-			}
-
-			return statistic;
 		}
 	}
 }
